requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
backports-zoneinfo==0.2.1
    # via
    #   django-celery-beat
    #   django-timezone-field
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   django-celery-beat
    #   django-celery-results
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via djangorestframework-stubs
coreschema==0.0.4
    # via coreapi
cryptography==39.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-rest-auth==2.2.3
    # via -r requirements.in
django==3.2.16
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-celery-beat
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-storages
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-jsonapi
    #   djangorestframework-simplejwt
django-allauth==0.50.0
    # via -r requirements.in
django-celery-beat==2.4.0
    # via -r requirements.in
django-celery-results==2.4.0
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
django-stubs==1.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-stubs
django-stubs-ext==0.7.0
    # via django-stubs
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-jsonapi
    #   djangorestframework-simplejwt
djangorestframework-camel-case==1.3.0
    # via -r requirements.in
djangorestframework-jsonapi==4.3.0
    # via -r requirements.in
djangorestframework-simplejwt==4.8.0
    # via -r requirements.in
djangorestframework-stubs==1.4.0
    # via -r requirements.in
idna==3.4
    # via requests
inflect==6.0.2
    # via -r requirements.in
inflection==0.5.1
    # via djangorestframework-jsonapi
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
kombu==5.2.4
    # via celery
markupsafe==2.1.2
    # via jinja2
mypy==1.0.1
    # via
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==1.0.0
    # via mypy
numpy==1.24.2
    # via
    #   -r requirements.in
    #   pandas
oauthlib==3.2.2
    # via requests-oauthlib
pandas==1.5.3
    # via -r requirements.in
prompt-toolkit==3.0.38
    # via click-repl
pycparser==2.21
    # via cffi
pydantic==1.10.5
    # via inflect
pyenchant==3.2.2
    # via -r requirements.in
pyjwt[crypto]==2.6.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   pandas
    #   python-crontab
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7.1
    # via
    #   celery
    #   django
    #   django-timezone-field
    #   pandas
requests==2.28.2
    # via
    #   coreapi
    #   django-allauth
    #   djangorestframework-stubs
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.3
    # via django
toml==0.10.2
    # via django-stubs
tomli==2.0.1
    # via mypy
types-pytz==2022.7.1.2
    # via django-stubs
types-pyyaml==6.0.12.8
    # via django-stubs
typing-extensions==4.5.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
    #   pydantic
tzdata==2022.7
    # via django-celery-beat
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.14
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
